+++ scurrypy/__init__.py
@@ -2,8 +2,10 @@
 
 from .client import Client
 from .intents import Intents
+from .bases.addon import Addon
 
 __all__ = [
     "Client",
-    "Intents"
+    "Intents",
+    "Addon"
 ]