+++ scurrypy/api/components/__init__.py
@@ -1,6 +1,9 @@
 # scurrypy/api/components
 
+from ...enums.components import ButtonStyle
 from .button import Button
+
+from ...enums.components import SeparatorType
 from .layout import (
     ActionRow,
     Section, 
@@ -13,6 +16,8 @@ from .layout import (
     Container,
     Label
 )
+
+from ...enums.components import TextInputStyle
 from .modal import (
     TextInput, 
     FileUpload, 
@@ -21,6 +26,8 @@ from .modal import (
     CheckboxGroup, 
     Checkbox
 )
+
+from ...enums.components import DefaultValueType
 from .select_menu import (
     SelectOption, 
     StringSelect, 
@@ -31,13 +38,16 @@ from .select_menu import (
     MentionableSelect, 
     ChannelSelect
 )
+
 from .factory import MessageComponentFactory
 
 __all__ = [
     "MessageComponentFactory",
     
+    "ButtonStyle",
     "Button",
 
+    "SeparatorType",
     "ActionRow",
     "Section", 
     "TextDisplay", 
@@ -49,6 +59,7 @@ __all__ = [
     "Container",
     "Label",
 
+    "TextInputStyle",
     "TextInput", 
     "FileUpload", 
     "ListOption", 
@@ -56,6 +67,7 @@ __all__ = [
     "CheckboxGroup", 
     "Checkbox",
 
+    "DefaultValueType",
     "SelectOption", 
     "StringSelect", 
     "DefaultValue",